refactored to avoid having certain tests being run unintentionally
