+++ test/utils/testUtils.ts
@@ -1,0 +1,5 @@
+import {BookState} from "../../src/renderer/reducers/bookReducer";
+
+export function createBookState(override: Partial<BookState> = {}): BookState {
+    return {isLoadingBook: false, fileName: null, bookWithMeta: null, ...override};
+}
